+++ packages/builder/src/index.ts
@@ -1,2 +1,0 @@
-export * as Vue2Builder from './vue2/index'
-export * as Vue3Builder from './vue3/index'